website/requirements.txt
Flask
Flask-Limiter
mysql-connector-python